refactor(debug): migrate debug dashboard to consume shared client primitives
