+++ client/debug/dashboard-types.ts
@@ -25,8 +25,22 @@ import type {
   Turn,
   Notification,
   ToolFailure,
-} from '../../src/debug/schemas.js'
-import type { GlobalStats, StatsWindow, SubjectStats } from '../../src/stats/types.js'
+  GlobalStats,
+  StatsWindow,
+  SubjectStats,
+  RecurringTask,
+  DeferredPrompt,
+  Memo,
+  IdentityMappingEntry,
+  AuthorizedGroupEntry,
+  BillingWindow,
+  BillingRoleTotals,
+  BillingSubject,
+  BillingRequestRow,
+  BillingDetail,
+  AdminLlmKeyState,
+  AdminLlmSnapshot,
+} from '../shared/api-types.js'
 
 export type {
   Fact,
@@ -53,108 +67,18 @@ export type {
   GlobalStats,
   StatsWindow,
   SubjectStats,
-}
-
-export type RecurringTask = {
-  id: string
-  userId: string
-  title: string
-  rrule: string | null
-  nextRun: string | null
-  enabled: boolean
-  lastRun: string | null
-}
-
-export type DeferredPrompt = {
-  id: string
-  createdByUserId: string
-  prompt: string
-  fireAt: string
-  rrule: string | null
-  status: string
-}
-
-export type Memo = {
-  id: string
-  userId: string
-  content: string
-  summary: string | null
-  tags: readonly string[]
-  status: string
-  createdAt: string
-  updatedAt: string
-}
-
-export type IdentityMappingEntry = {
-  userId: string
-  provider: string
-  providerUserId: string | null
-  providerUserLogin: string | null
-  displayName: string | null
-}
-
-export type AuthorizedGroupEntry = {
-  group_id: string
-  added_by: string
-  added_at: string
-}
-
-export type BillingWindow = '24h' | '7d' | '30d' | 'all'
-
-export type BillingRoleTotals = {
-  inputTokens: number
-  outputTokens: number
-  calls: number
-}
-
-export type BillingSubject = {
-  storageContextId: string
-  contextType: 'dm' | 'group'
-  displayName: string | null
-  totals: {
-    main: BillingRoleTotals
-    small: BillingRoleTotals
-    embedding: BillingRoleTotals
-  }
-  toolCalls: number
-  lastActiveAt: number
-}
-
-export type BillingRequestRow = {
-  eventId: string
-  occurredAt: number
-  turnId: string | null
-  chatUserId: string
-  model: string
-  modelRole: 'main' | 'small' | 'embedding'
-  inputTokens: number | null
-  outputTokens: number | null
-  stepCount: number
-  toolCallCount: number
-  messageCount: number
-  durationMs: number
-  finishReason: string | null
-  error: string | null
-}
-
-export type BillingDetail = {
-  subject: BillingSubject
-  requests: readonly BillingRequestRow[]
-  truncated: boolean
-}
-
-export type AdminLlmKeyState = {
-  value: string | null
-  updatedAt: number | null
-  updatedBy: string | null
-}
-
-export type AdminLlmSnapshot = {
-  llm_apikey: AdminLlmKeyState
-  llm_baseurl: AdminLlmKeyState
-  main_model: AdminLlmKeyState
-  small_model: AdminLlmKeyState
-  embedding_model: AdminLlmKeyState
+  RecurringTask,
+  DeferredPrompt,
+  Memo,
+  IdentityMappingEntry,
+  AuthorizedGroupEntry,
+  BillingWindow,
+  BillingRoleTotals,
+  BillingSubject,
+  BillingRequestRow,
+  BillingDetail,
+  AdminLlmKeyState,
+  AdminLlmSnapshot,
 }
 
 /**
