--- client/debug/helpers.ts
@@ -1,52 +1,0 @@
-// SPDX-License-Identifier: BUSL-1.1
-// Copyright (c) 2026 Dmitriy Lazarev
-// Use of this software is governed by the Business Source License 1.1.
-// See LICENSE in the project root for details.
-
-const LEVEL_NAMES: Record<number, string> = {
-  10: 'trace',
-  20: 'debug',
-  30: 'info',
-  40: 'warn',
-  50: 'error',
-  60: 'fatal',
-}
-
-export function levelName(level: number): string {
-  return LEVEL_NAMES[level] ?? `L${level}`
-}
-
-export function levelClass(level: number): string {
-  if (level >= 50) return 'log-error'
-  if (level >= 40) return 'log-warn'
-  if (level >= 30) return 'log-info'
-  return 'log-debug'
-}
-
-export function formatTime(ts: number | string): string {
-  const d = typeof ts === 'string' ? new Date(ts) : new Date(ts)
-  return d.toLocaleTimeString('en-GB', {
-    hour12: false,
-    hour: '2-digit',
-    minute: '2-digit',
-    second: '2-digit',
-  })
-}
-
-export function formatUptime(startedAt: number): string {
-  const ms = Date.now() - startedAt
-  const s = Math.floor(ms / 1000)
-  const h = Math.floor(s / 3600)
-  const m = Math.floor((s % 3600) / 60)
-  if (h > 0) return `${h}h${m}m`
-  return `${m}m${s % 60}s`
-}
-
-export function formatTokens(n: number): string {
-  if (n >= 1000) return `${(n / 1000).toFixed(1)}k`
-  return String(n)
-}
-
-export function escapeHtml(str: string): string {
-  return str.replace(/&/gu, '&amp;').replace(/</gu, '&lt;').replace(/>/gu, '&gt;').replace(/"/gu, '&quot;')
-}
